# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: NXG Ecommerce API
  description: 'This is a NXG Ecommerce module. You can find out more about NXG Ecommerce at [http://nxgecommerce.in](http://nxgecommerce.io) or on [www.nxgecommerce.in, #nxgecommerce].'
  termsOfService: http://nxgecommerce/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://localhost:8080/nxgecommerce/api/v1
  - url: http://localhost:8080/nxgecommerce/api/v1
...
paths:
  /category:
    post:
      tags:
        - Category Information
      summary: Adding a new category
      requestBody:
        description: category details that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Categories'
        required: true
      responses:
        '201':
          description: Successful
          content: {}      
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}          
        '502':
          description: Bad Gatway
          content: {}               
      x-codegen-request-body-name: body
    get:
      tags:
        - Category Information
      summary:  Getting all categories details
      responses:
        '200':
          description:  A list of all categories details
          content:
            applicantion/json:
              schema:
                $ref: '#/components/schemas/Categories'
        '201':
          description: Successful
          content: {}
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}
        '502':
          description: Bad Gatway
      x-codegen-request-body-name: body
  /category/{id}:
    get:
      tags:
        - Category Information
      summary: Getting Category details by id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: A list of a category details
          content:
            applicantion/json:
              schema:
                $ref: '#/components/schemas/Categories'
        '201':
          description: Successful
          content: {}
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}
        '502':
          description: Bad Gatway
      x-codegen-request-body-name: body   
    put:
        tags:
          - Category Information
        summary: Editing category details
        parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        requestBody:
          description: Category details that need to be added
          content:
            applicantion/json:
              schema:
                $ref: '#/components/schemas/Categories'
          required: true      
        responses:
          '200':
            description: A list of a category details
            content:
              applicantion/json:
                schema:
                  $ref: '#/components/schemas/Categories'
          '201':
            description: Successful
            content: {}
          '400':
            description: Bad request input
            content: {}
          '404':
            description: Data not found
            content: {}
          '500':
            description: Invalid input
            content: {}
          '502':
            description: Bad Gatway
        x-codegen-request-body-name: body
    delete:
      tags:
        - Category Information
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Successful
          content: {}      
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}          
        '502':
          description: Bad Gatway
          content: {}               
      x-codegen-request-body-name: body   
...
components:
  schemas:
    Categories:
      type: object
      required:
        - name
        - position
        - image
        - categoryHandel
      properties:
        id:
          type: integer
          maxLength: 11
        name:
          type: string
          maxLength: 200
        position:
          type: integer
          maxLength: 4
        image:
          type: string
        categoryHandel:
          type: string
          maxLength: 200
...